+++ js/util.js
@@ -1,0 +1,11 @@
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (arr) => arr[getRandomInteger(0, arr.length - 1)];
+
+export {getRandomInteger, getRandomArrayElement};
